giga progress 7

+++ src/main/java/org/example/resources/MovieVersionResource.java
@@ -3,8 +3,7 @@ package org.example.resources;
 
 import jakarta.ws.rs.*;
 import org.example.objects.LowRes;
-import org.example.objects.MovieVersion;
-import org.example.objects.StreamThread;
+import org.example.objects.M3u8Thread;
 import org.example.repositories.MovieRepository;
 import org.example.repositories.MovieVersionRepository;
 
@@ -40,7 +39,10 @@ public class MovieVersionResource {
             return Response.status(Response.Status.BAD_REQUEST).entity("Invalid ID").build();
         }
         String movieVersionPath = MovieVersionRepository.getInstance().getMovieVersion(id).getMovieLink();
-        new File(movieVersionPath).delete();
+        File versionToDelete = new File(findFolder(movieVersionPath));
+        if (!deleteDirectory(versionToDelete)) {
+            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+        }
         MovieVersionRepository.getInstance().deleteMovieVersion(id);
         return Response.ok().entity("Movie Version deleted").build();
     }
@@ -59,39 +61,27 @@ public class MovieVersionResource {
         if (ext.isEmpty()) {
             return Response.status(415).entity("need format").build();
         }
-        String ProjPath = System.getProperty("user.dir");
-        String fileLocation = ProjPath + "/src/main/resources/videos/High/" + formData.getFileName();
-        String lowResFileLocation = ProjPath + "/src/main/resources/videos/Low/" + formData.getFileName();
+        String ProjPath = "/home/miguel/movieFiles/";
+        File highPath = new File(ProjPath + "high/" + removeExt(formData.getFileName()) + getExt(formData.getFileName()));
+        File lowPath = new File(ProjPath + "low/" + removeExt(formData.getFileName()) + getExt(formData.getFileName()));
+        highPath.mkdir();
+        lowPath.mkdir();
+        String fileLocation = highPath + "/" + formData.getFileName();
+        String lowResFileLocation = lowPath + "/" + formData.getFileName();
         try {
-            saveFile(is, fileLocation, lowResFileLocation, Integer.parseInt(movieId), ext);
-            LowRes lowRes = new LowRes(formData.getFileName(), ext, ProjPath);
+            saveFile(is, fileLocation, Integer.parseInt(movieId), formData.getFileName());
+            LowRes lowRes = new LowRes(fileLocation, ext, lowResFileLocation, Integer.parseInt(movieId), formData.getFileName());
             lowRes.start();
-            return Response.status(Response.Status.OK).entity("completed").build();
+            return Response.status(Response.Status.OK).entity("completed base upload").build();
         } catch (IOException e) {
             return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
         }
     }
 
-    @POST
-    @Path("/stream")
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.TEXT_PLAIN)
-    public Response movieStream(MovieVersion movieVersion) {
-        String moviePath = movieVersion.getMovieLink();
-        int movieId = movieVersion.getMovieId();
-        String movieRes = movieVersion.getMovieResolution();
-        String movieFormat = movieVersion.getMovieFormat();
-        String URL;
-        File f1 = new File("/tmp/hls/" + movieId + movieRes + movieFormat + "-0.ts");
-        URL = "http://192.168.1.103:1234/" + movieId + movieRes + movieFormat + ".m3u8";
-        if (f1.exists()) {
-            return Response.status(200).entity(URL).build();
-        }
 
-        StreamThread streamThread = new StreamThread(movieId, moviePath, movieRes, movieFormat);
-        streamThread.start();
-
-        return Response.status(200).entity(URL).build();
+    private String removeExt(String filename) {
+        int lasIndexOf = filename.lastIndexOf(".");
+        return filename.substring(0, lasIndexOf);
     }
 
     private String getExt(String fileName){
@@ -102,7 +92,24 @@ public class MovieVersionResource {
         return fileName.substring(lastIndexOf).substring(1);
     }
 
-    private void saveFile(InputStream is, String fileLocation, String lowResFileLocation, int movieId, String format) throws FileNotFoundException {
+    private String findFolder(String path) {
+        int lastIndexOf = path.lastIndexOf("/");
+        int slashCount = 0;
+        int startIndex = 0;
+        for (int i = 0; i < path.length(); i++) {
+            if (path.charAt(i) == ('/')) {
+                slashCount++;
+                if (slashCount == 3) {
+                    startIndex = i;
+                    break;
+                }
+            }
+        }
+
+        return "/home/miguel/movieFiles" + path.substring(startIndex, lastIndexOf);
+    }
+
+    private void saveFile(InputStream is, String fileLocation, int movieId, String filename) throws FileNotFoundException {
         try {
             OutputStream os = new FileOutputStream(fileLocation);
             byte[] buffer = new byte[256];
@@ -112,8 +119,20 @@ public class MovieVersionResource {
             }
             os.close();
             is.close();
-            MovieVersionRepository.getInstance().insertMovieVersion(new MovieVersion(movieId, format, "high", fileLocation));
-            MovieVersionRepository.getInstance().insertMovieVersion(new MovieVersion(movieId, format, "low", lowResFileLocation));
-        } catch (IOException | SQLException ignored) {}
+
+            M3u8Thread m3u8Thread = new M3u8Thread(fileLocation, getExt(fileLocation), movieId, "high", filename);
+            m3u8Thread.start();
+        } catch (IOException ignored) {}
+    }
+
+    boolean deleteDirectory(File directoryToBeDeleted) {
+        File[] allContents = directoryToBeDeleted.listFiles();
+        if (allContents != null) {
+            for (File file : allContents) {
+                deleteDirectory(file);
+            }
+        }
+        return directoryToBeDeleted.delete();
     }
 }
+
